adds initial setup and basic testing in CompletedProfileViewController.swift for new OwnerFirestore object creation and save to tests in Firebase via browser dashboard view and adds a SnapshotListener to same VC to pulldown some of the newly saved data to the app and print to console also it is an owner object tested via student profile creation in app to aoid adding too many owner models in testing sparing any unnecessary confusion
