+++ Academia/Controllers/View Controllers/LoginViewController.swift
@@ -50,10 +50,10 @@ class LoginViewController: UIViewController {
     var activeOwnerDelegate: ActiveOwnerDelegate?
     var activeStudentDelegate: ActiveStudentDelegate?
     
-    // Firebase Firestore properties
-    var docRef: DocumentReference!
-    var firestoreTestListener: ListenerRegistration!
-    var db: Firestore!
+//    // Firebase Firestore properties
+//    var docRef: DocumentReference!
+//    var firestoreTestListener: ListenerRegistration!
+//    var db: Firestore!
     
     
     
@@ -75,37 +75,35 @@ class LoginViewController: UIViewController {
         passwordTextField.isSecureTextEntry = true
         
         
-        // FIREBASE FIRESTORE docRef listeners
-        
-        firestoreTestListener = docRef.addSnapshotListener { (docSnapshot, error) in
-            
-            guard let docSnapshot = docSnapshot, docSnapshot.exists else {
-                print("ERROR: no docSnapshot in LoginVC.swift -> loginButtonTapped - line 134.")
-                return
-            }
-            if let error = error {
-                print("ERROR: error: \(error.localizedDescription) occurred while trying to get docSnapshot in LoginVC.swift -> loginButtonTapped - line 138. ")
-            } else {
-                print("successful docSnapshot retrieval!")
-                if let myData = docSnapshot.data() {
-                    
-                    let usernameTextInput = myData["username"] as? String ?? "username fail"
-                    let passwordTextInput = myData["password"] as? String ?? "password fail"
-                    
-                    print("docID: \(docSnapshot.documentID)\nusername: \(usernameTextInput)\npassword: \(passwordTextInput)")
-                }
-            }
-        }
-        
-        
+//        // FIREBASE FIRESTORE docRef listeners
+//
+//        firestoreTestListener = docRef.addSnapshotListener { (docSnapshot, error) in
+//
+//            guard let docSnapshot = docSnapshot, docSnapshot.exists else {
+//                print("ERROR: no docSnapshot in LoginVC.swift -> loginButtonTapped - line 134.")
+//                return
+//            }
+//            if let error = error {
+//                print("ERROR: error: \(error.localizedDescription) occurred while trying to get docSnapshot in LoginVC.swift -> loginButtonTapped - line 138. ")
+//            } else {
+//                print("successful docSnapshot retrieval for OwnerFirestore model!")
+//                if let myData = docSnapshot.data() {
+//
+//                    let usernameTextInput = myData["username"] as? String ?? "username fail"
+//                    let passwordTextInput = myData["password"] as? String ?? "password fail"
+//
+//                    print("docID: \(docSnapshot.documentID)\nusername: \(usernameTextInput)\npassword: \(passwordTextInput)")
+//                }
+//            }
+//        }
     }
     
     override func viewWillDisappear(_ animated: Bool) {
         
         unsubscribeToKeyboardNotifications()
         
-        // FIREBASE FIRESTORE remove the test listener to avoid reference cycle
-        firestoreTestListener.remove()
+//        // FIREBASE FIRESTORE remove the test listener to avoid reference cycle
+//        firestoreTestListener.remove()
     }
 
     override func viewDidLoad() {
@@ -117,9 +115,9 @@ class LoginViewController: UIViewController {
         usernameTextField.delegate = self
         passwordTextField.delegate = self
         
-        // Firestore Test properties setup
-        docRef = Firestore.firestore().collection("tests").document("level1")
-        db = Firestore.firestore()
+//        // Firestore Test properties setup
+//        docRef = Firestore.firestore().collection("tests").document("level1")
+//        db = Firestore.firestore()
     }
 
     // MARK: - Actions
@@ -144,39 +142,39 @@ class LoginViewController: UIViewController {
         }
         
         
-        // FIREBASE FIRESTORE TESTING
-        guard let usernameText = usernameTextField.text, !usernameText.isEmpty else {
-            print("ERROR: no valid text input in the usernameTextField in LoginVC.swift -> loginButtonTapped - line 108.")
-            return
-        }
-        guard let passwordText = passwordTextField.text, !passwordText.isEmpty else {
-            print("ERROR: no valid text input in the passwordTextField in LoginVC.swift -> loginButtonTapped - line 112.")
-            return
-        }
-        
-        let dataToSave: [String : Any] = ["username" : usernameText, "password" : passwordText]
-        
-        docRef.setData(dataToSave) { (error) in
-            if let error = error {
-                print("ERROR: error: \(error.localizedDescription) occurred while trying to save to Firebase Firestore in LoginVC.swift -> loginButtonTapped - line 125. ")
-            } else {
-                print("Data successfully saved to Firebase Firestore")
-            }
-        }
-        
-        // test for new testModel object creation within Firestore using the username and password checks in lines 143 and 147 above, respectively.
-        let test = TestModel(username: usernameText, password: passwordText)
-        
-        let testModelRef = self.db.collection("tests")
-        
-        testModelRef.document(/*usernameText*/).setData(test.dictionary) { (error) in
-            // ^^^ NOTE: the document id can be created and set to my specifications as it is currently being set in the .document(usernameText) portion of the method call above.  an option would be to have the id as its own key:value pair in the TestModel dictionary property itself, and then access it via the dictionary key and set it in below in the place of the current 'usernameText' property
-            if let error = error {
-                print("ERROR: error: \(error.localizedDescription) occurred while trying to save test.dictionary to Firestore in LoginVC.swift -> loginButtonTapped - line 173. ")
-            } else {
-                print("test dictionary data successfully saved to Firestore document in tests collection")
-            }
-        }
+//        // FIREBASE FIRESTORE TESTING
+//        guard let usernameText = usernameTextField.text, !usernameText.isEmpty else {
+//            print("ERROR: no valid text input in the usernameTextField in LoginVC.swift -> loginButtonTapped - line 108.")
+//            return
+//        }
+//        guard let passwordText = passwordTextField.text, !passwordText.isEmpty else {
+//            print("ERROR: no valid text input in the passwordTextField in LoginVC.swift -> loginButtonTapped - line 112.")
+//            return
+//        }
+//
+//        let dataToSave: [String : Any] = ["username" : usernameText, "password" : passwordText]
+//
+//        docRef.setData(dataToSave) { (error) in
+//            if let error = error {
+//                print("ERROR: error: \(error.localizedDescription) occurred while trying to save to Firebase Firestore in LoginVC.swift -> loginButtonTapped - line 125. ")
+//            } else {
+//                print("Data successfully saved to Firebase Firestore")
+//            }
+//        }
+//
+//        // test for new testModel object creation within Firestore using the username and password checks in lines 143 and 147 above, respectively.
+//        let test = TestModel(username: usernameText, password: passwordText)
+//
+//        let testModelRef = self.db.collection("tests")
+//
+//        testModelRef.document(/*usernameText*/).setData(test.dictionary) { (error) in
+//            // ^^^ NOTE: the document id can be created and set to my specifications as it is currently being set in the .document(usernameText) portion of the method call above.  an option would be to have the id as its own key:value pair in the TestModel dictionary property itself, and then access it via the dictionary key and set it in below in the place of the current 'usernameText' property
+//            if let error = error {
+//                print("ERROR: error: \(error.localizedDescription) occurred while trying to save test.dictionary to Firestore in LoginVC.swift -> loginButtonTapped - line 173. ")
+//            } else {
+//                print("test dictionary data successfully saved to Firestore document in tests collection")
+//            }
+//        }
         
         
         // check for required information being left blank by user
