+++ Academia/Controllers/View Controllers/User Onboarding/CompletedProfileViewController.swift
@@ -8,6 +8,8 @@
 
 import UIKit
 import CoreData
+import FirebaseCore
+import FirebaseFirestore
 
 class CompletedProfileViewController: UIViewController {
     
@@ -76,8 +78,45 @@ class CompletedProfileViewController: UIViewController {
     @IBOutlet weak var emergencyContactRelationshipLabelOutlet: UILabel!
     @IBOutlet weak var emergencyContactPhoneLabelOutlet: UILabel!
     
+    // Firebase Firestore properties
+    var docRef: DocumentReference!
+    var firestoreTestListener: ListenerRegistration!
+    var db: Firestore!
+    
     
     // MARK: - ViewController Lifecycle Functions
+    
+    override func viewWillAppear(_ animated: Bool) {
+        // FIREBASE FIRESTORE docRef listeners
+        
+        firestoreTestListener = docRef.addSnapshotListener { (docSnapshot, error) in
+            
+            guard let docSnapshot = docSnapshot, docSnapshot.exists else {
+                print("ERROR: no docSnapshot in LoginVC.swift -> loginButtonTapped - line 134.")
+                return
+            }
+            if let error = error {
+                print("ERROR: error: \(error.localizedDescription) occurred while trying to get docSnapshot in LoginVC.swift -> loginButtonTapped - line 138. ")
+            } else {
+                print("successful docSnapshot retrieval!")
+                if let myData = docSnapshot.data() {
+                    
+                    let usernameTextInput = myData["username"] as? String ?? "username fail"
+                    let passwordTextInput = myData["password"] as? String ?? "password fail"
+                    let firstNameTextInput = myData["firstName"] as? String ?? "no first name"
+                    let lastNameTextInput = myData["lastName"] as? String ?? "no last name"
+                    
+                    print("docID: \(docSnapshot.documentID)\nusername: \(usernameTextInput)\npassword: \(passwordTextInput)\nfirstName: \(firstNameTextInput)\nlastName: \(lastNameTextInput)")
+                }
+            }
+        }
+    }
+    
+    override func viewWillDisappear(_ animated: Bool) {
+        
+        // FIREBASE FIRESTORE remove the test listener to avoid reference cycle
+        firestoreTestListener.remove()
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -99,12 +138,65 @@ class CompletedProfileViewController: UIViewController {
                 }
             }
         }
+        
+        // Firestore Test properties setup
+        docRef = Firestore.firestore().collection("tests").document("ownerTest")
+        db = Firestore.firestore()
     }
     
     
     // MARK: - Actions
     
     @IBAction func createAccountButtonTapped(_ sender: DesignableButton) {
+    
+        // FIREBASE FIRESTORE TESTING
+        guard let username = username else { print("fail username"); return }
+        guard let password = password else { print("fail password"); return }
+        
+        guard let firstName = firstName else { print("fail firtsName"); return }
+        guard let lastName = lastName else { print("fail lastName"); return }
+//        guard let profilePic = profilePic else { print("fail profilePic"); return }
+        
+        //        guard let birthdate = birthdate else { print("fail birthdate"); return }
+        
+//        guard let beltCD = beltCD else { print("fail beltCD"); return }
+        
+//        guard let addressCD = addressCD else { print("fail addressCD"); return }
+        
+        guard let phone = phone else { print("fail phone"); return }
+        guard let email = email else { print("fail email"); return }
+        
+//        guard let emergencyContactCD = emergencyContactCD else { print("fail emergencyContactCD"); return }
+        
+        let mobileNumber = mobile ?? ""
+        
+//         convert profilePic to Data
+//        guard let profilePicData = profilePic.jpegData(compressionQuality: 1) else { print("fail profilePicData"); return }
+        
+        let dataToSave: [String : Any] = ["username" : username, "password" : password, "firstName" : firstName, "lastName" : lastName, "phone" : phone, "email" : email, "mobile" : mobileNumber]
+        
+        docRef.setData(dataToSave) { (error) in
+            if let error = error {
+                print("ERROR: error: \(error.localizedDescription) occurred while trying to save to Firebase Firestore in LoginVC.swift -> loginButtonTapped - line 180. ")
+            } else {
+                print("Data successfully saved to Firebase Firestore")
+            }
+        }
+        
+        // test for new testModel object creation within Firestore using the username and password checks in lines 143 and 147 above, respectively.
+        let test = OwnerFirestore(mostRecentPromotion: Date(), profilePic: "URL I haven't set up yet", username: username, password: password, firstName: firstName, lastName: lastName, phone: phone, mobile: mobileNumber, email: email)
+        
+        let testModelRef = self.db.collection("tests")
+        
+        testModelRef.document(/*usernameText*/).setData(test.dictionary) { (error) in
+            // ^^^ NOTE: the document id can be created and set to my specifications as it is currently being set in the .document(usernameText) portion of the method call above.  an option would be to have the id as its own key:value pair in the TestModel dictionary property itself, and then access it via the dictionary key and set it in below in the place of the current 'usernameText' property
+            if let error = error {
+                print("ERROR: error: \(error.localizedDescription) occurred while trying to save test.dictionary to Firestore in LoginVC.swift -> loginButtonTapped - line 194. ")
+            } else {
+                print("test dictionary data successfully saved to Firestore document in tests collection")
+            }
+        }
+        
         
         // create and save new user account to CoreData
         createAndSaveNewUser()
